Adds sample notebook
Please make sure that you have a Github account. If you do not, head over to github.com and click on the Sign up button on the top right corner and follow the instructions to create your account.
Please note that the steps and instructions described in this post are relevant to a Windows computer, so if you are using a Mac computer, please check the Mac version of this post.